pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.12",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
    "requests>=2.32.3",
    "telegram>=0.0.1",
    "telethon>=1.38.1",
    "twilio>=9.4.5",
]
